Changed PDF417 to generate barcodes that fit within the specified height/width if possible.
Added PDF417AspectRatio, when set, changes the height of a row from the default of 4 to specified value.  AUTO will use the largest value possible that fits.

Added PDF417ErrorCorrection.AUTO, when set, uses the minimumErrorCorrectionLevel for the specified text per ISO/IEC 15438:2001(E).

Minor optimization for BarcodeMatrix, now generates a smaller matrix for PDF417.compact instead of 34px of white space on the right.